pyproject.toml
[tool.poetry]
name = "action-shot-extractor"
version = "0.1.0"
description = "AI-powered action shot extractor for sports and dynamic video content."
authors = ["Justin Robinson <justin.g.robinson@gmail.com>"]
readme = "README.md"
packages = [{ include = "action_shot_extractor", from = "src" }]

[tool.poetry.dependencies]
python = "^3.13"
click = "^8.1.7"
ultralytics = "^8.0.0"
torch = "^2.2.0"
torchvision = "^0.17.0"
torchaudio = "^2.2.0"
opencv-python = "^4.8.0"
pillow = "^10.0.0"
numpy = "^1.26.0"
tqdm = "^4.66.0"
transformers = "^4.37.0"
huggingface_hub = "^0.19.0"
scikit-image = "^0.22.0"
colorama = "^0.4.6"

[tool.poetry.dev-dependencies]
pytest = "^8.0.0"

[tool.poetry.scripts]
action-shot-extractor = "action_shot_extractor.cli:main"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
